Restructure project and exports to support CommonJS and ESModules
