+++ package.json
@@ -6,22 +6,15 @@
     "/lib",
     "/src"
   ],
+  "type": "commonjs",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
   "exports": {
     ".": {
       "default": "./lib/index.js",
       "types": "./lib/index.d.ts"
-    },
-    "./parse": {
-      "default": "./lib/parse.js",
-      "types": "./lib/parse.d.ts"
-    },
-    "./io": {
-      "default": "./lib/io.js",
-      "types": "./lib/io.d.ts"
     }
   },
-  "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
   "scripts": {
     "build": "rimraf lib && tsc",
     "test": "echo \"Error: no test specified\" && exit 1"
