--- src/types.ts
@@ -1,5 +1,0 @@
-export interface Property {
-    name: string
-    params: string[]
-    value: string
-}
